check_can_submit help function for pre-propose-base
