+++ contracts/pre-propose/cwd-pre-propose-approval-single/Cargo.toml
@@ -21,7 +21,6 @@ cosmwasm-schema = { workspace = true }
 cw-storage-plus = { workspace = true }
 cw2 = { workspace = true }
 cw-paginate = { workspace = true }
-cwd-interface = { workspace = true }
 cwd-pre-propose-base = { workspace = true }
 cwd-proposal-single = { workspace = true, features = ["library"] }
 cwd-voting = { workspace = true }
